requirements.txt
google-cloud-aiplatform
google-cloud-discoveryengine
google-generativeai
streamlit
python-dotenv